move defaultDecoder & defaultResolver to options.go
