+++ config/options.go
@@ -1,6 +1,11 @@
 package config
 
 import (
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/go-kratos/kratos/v2/encoding"
 	"github.com/go-kratos/kratos/v2/log"
 )
 
@@ -46,4 +51,55 @@ func WithLogger(l log.Logger) Option {
 	return func(o *options) {
 		o.logger = l
 	}
+}
+
+// defaultDecoder decode config from source KeyValue
+// to target map[string]interface{} using src.Format codec.
+func defaultDecoder(src *KeyValue, target map[string]interface{}) error {
+	if src.Format == "" {
+		target[src.Key] = src.Value
+		return nil
+	}
+	if codec := encoding.GetCodec(src.Format); codec != nil {
+		return codec.Unmarshal(src.Value, &target)
+	}
+	return fmt.Errorf("unsupported key: %s format: %s", src.Key, src.Format)
+}
+
+// defaultResolver resolve placeholder in map value,
+// placeholder format in ${key:default} or $key.
+func defaultResolver(input map[string]interface{}) error {
+	mapper := func(name string) string {
+		args := strings.Split(strings.TrimSpace(name), ":")
+		if v, has := readValue(input, args[0]); has {
+			s, _ := v.String()
+			return s
+		} else if len(args) > 1 { // default value
+			return args[1]
+		}
+		return ""
+	}
+
+	var resolve func(map[string]interface{}) error
+	resolve = func(sub map[string]interface{}) error {
+		for k, v := range sub {
+			switch vt := v.(type) {
+			case string:
+				sub[k] = os.Expand(vt, mapper)
+			case map[string]interface{}:
+				if err := resolve(vt); err != nil {
+					return err
+				}
+			case []interface{}:
+				for i, iface := range vt {
+					if s, ok := iface.(string); ok {
+						vt[i] = os.Expand(s, mapper)
+					}
+				}
+				sub[k] = vt
+			}
+		}
+		return nil
+	}
+	return resolve(input)
 }
